pyproject.toml
[tool.poetry]
name = "dcbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
PyNaCl = "^1.4.0"
replit = "^3.2.4"
pydub = "^0.25.1"
gTTS = "^2.2.3"
youtube_dl = "^2021.12.17"
youtube-search = "^2.1.0"
"discord.py" = "2.2.2"
openai = "^0.27.0"
asgiref = "^3.6.0"
pytz = "^2022.7.1"
constant = "^0.0.4"
quart = "0.18.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
